requirements.txt
djangorestframework
psycopg2-binary
twilio
Pillow
python-decouple
drf-yasg
djangorestframework-simplejwt
redis
celery
django-filter
python-slugify
